requirements.txt
eth-brownie>=1.14.6,<2.0.0
pytest>=6.2.3
brownie-token-tester>=0.2.2
black
isort
mypy